requirements.txt
numpy~=1.23.1
torch~=1.12.1
timm~=0.6.7
torchvision~=0.13.1
PyYAML~=6.0